1.Z.2
Possible backbone for the graphics part of the project. Lots can be changed to fit what we need and features can be added. I plan to add a way to navigate the text boxes by using the keyboard and not just the mouse.

The way things are setup are as follows:
The teams are part of a 2D array where  the first index determines red or green team.
Each text box has its own string to display and its own flag for changing the color if interacted with.
Each text box will match size with the current largest text box.
Clicking the window X, the quit button, or pressing F4 will close the window.